library_web/requirements.txt
flask==3.1.1
mysql-connector-python==9.4.0
python-dotenv==1.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.2.1
itsdangerous==2.2.0
MarkupSafe==3.0.2
blinker==1.9.0
colorama==0.4.6
